:sparkles: Support for configuration files
There are two locations for the configuration file:

- System-Defined: /usr/lib/system76-power/config.toml
- User-Defined: /etc/system76-power/config.toml

Values which are set in the user-defined configuration file are
overlayed onto the system-defined configuration parameters. This
eans that the system config is used as a template for setting
the default values, and then any values stored in the user-defined
onfiguration file will overwrite the default values.

Maintainers are therefore required to provide a system config file
ith their packages. Included is the default configuration used by
Pop!_OS.