feat: Add dbus API and CLI for battery charge thresholds
The dbus API authenticates with Polkit, which should prompt for an
administrator password, or run without prompt when the client in root.

Thresholds can be displayed in the CLI with `system76-power
charge-thresholds`, and set with `system76-power charge-thresholds
<start> <end>`.

This is currently limited to System76 hardware with open firmware, to
avoid worrying for now about things like systems with multiple
batteries, only having an end thresholds, and other differences between
implementations.