graphics: Get NVIDIA GPU features without driver loaded
Store the device ID at PCI enumeration so it does not need to be fetched
from sysfs again later.

Search /usr/share/doc for the NVIDIA driver folder instead of getting
the version of the loaded driver.

Allows getting the dGPU features when the driver is unloaded and the PCI
device has been removed from the bus.

Signed-off-by: Tim Crawford <tcrawford@system76.com>