fan: Simplify and clarify the interpolation functions
First of all, this clarifies what kind of interpolation is done in
"interpolate_duties" by renaming it to "linear_interpolation_duty". This
function mathematically handles the cases where the point to be fit is
one of the boundaries so there is no need for extra logic of that in
"get_duty_between_points".

This change also includes some minor rewordings and formatting.

Signed-off-by: Andrei Gherzan <andrei@gherzan.com>