Add NVME temperatures in fan speed calculation
Implement a second fan curve for NVME temperatures, and take the maximum
fan speed from the CPU duty curve and the NVME duty curve and use that
to set fan speed.

The NVME fan curve is hardcoded to a single table assuming 70 degrees
Celsius is the hottest allowable temperature (i.e., the fans should
already be at 100% before 70 degrees is reached).

Signed-off-by: Erich Ritz <erich.public@protonmail.com>